Change building abbreviations to upper case
ECSOPS-77

We store the building foreign key in banner using the abbreviation.
In Banner we use uppercase for the building abbreviation, so these
vcf files must match that.

The dining locations script has a manual static map of locations
because from dining we don't have a consistent name / map to locations.
The data we get from dining has slightly different building names or
abbreviations which make it not possible to map.