Don't serialize twice
ECSPCS-311

Serializing would throw an exception the first time because vobject
expects N fields to be vcard.Name objects. The serialization code would
set isNative on the field and then attempt to convert the Name object to
a string, which fails because it is already a string. Catching the
exception leaves isNative set, so the second time we serialize it
doesn't try to convert the object again.

Instead of all that, just set isNative ourselves.

Note that the resulting vCard is not strictly conforming, since the
N field is supposed to have 5 semicolon-separated components.[1]

[1] https://tools.ietf.org/html/rfc6350#section-6.2.2